test(domain): split PositionsViewInputTests into per-type files
PositionsViewInputTests.swift previously hosted three @Suite structs
(ValuedPositionTests, HistoricalValueSeriesTests, PositionsViewInputTests),
violating CODE_GUIDE §2 "one primary type per file, filename matches".
The size-threshold exemption for tests doesn't override the naming
discipline. Split into three files named after each primary type.
